BrandOtpOfficial/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pymongo
requests
PyJWT
passlib[bcrypt]
python-multipart
starlette
cryptography
dnspython
email-validator
bson
typing-extensions
Werkzeug
watchfiles
websockets
PyYAML
pylance
httpx
bcrypt
dnspython
Jinja2
